requirements.txt
black==23.11.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.3.2
databind.core==4.4.2
databind.json==4.4.2
Deprecated==1.2.14
docspec==2.2.1
docspec-python==2.2.1
docstring-parser==0.11
exceptiongroup==1.2.0
falcon==2.0.0
flake8==6.1.0
hug==2.6.1
idna==3.6
importlib-metadata==7.0.0
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
load-dotenv==0.1.0
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.7.1
mypy-extensions==1.0.0
nr-date==2.1.0
nr-stream==1.1.5
nr.util==0.8.12
packaging==23.2
pathspec==0.11.2
pendulum==2.1.2
platformdirs==4.1.0
pluggy==1.3.0
pycodestyle==2.11.1
pydoc-markdown==4.8.2
pyflakes==3.1.0
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytzdata==2020.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
testcase==0.1.0
tomli==2.0.1
tomli_w==1.0.0
typeapi==2.1.1
types-requests==2.31.0.10
types-waitress==2.1.4.9
typing_extensions==4.6.3
urllib3==2.1.0
waitress==2.1.2
watchdog==3.0.0
wrapt==1.16.0
yapf==0.40.2
zipp==3.17.0
